sample code for zephir cluster lookup
